New: Select › All Artboards in Page
This command selects all Artboards in current page
so you can align them, make a grid, or edit them
to your hearts content.

When you run the command, it will select the
Artboards, even if they don't appear highlighted
on the left layer list. This is a known bug, and I
will fix it when I find the time :)

This closes #16